Completed "IPO" question
Used a custom Project class to create projects. And Used a Priority queue to choose projects with highest profit gain.
The projects are only added to this priority queue iff the project capital is less than the total profit so far.
This is done by sorting all projects in ascending order by their capital value.